+++ backend/mysite/Trinity_Project/azure_file_share.py
@@ -1,159 +1,130 @@
-import os
-from azure.storage.fileshare import ShareServiceClient, ShareDirectoryClient
+from azure.storage.fileshare import ShareServiceClient, ShareClient, ShareDirectoryClient
 from azure.core.exceptions import ResourceExistsError
 from django.conf import settings
 
-def get_share_service_client():
-    connection_string = settings.AZURE_FILE_SHARE_CONNECTION_STRING
-    share_service_client = ShareServiceClient.from_connection_string(connection_string)
-    return share_service_client
-
-def create_folder_in_file_share(folder_name):
-    # Get the share service client
-    share_service_client = get_share_service_client()
-
-    # Get the file share client
-    file_share_name = settings.AZURE_FILE_SHARE_NAME
-    file_share_client = share_service_client.get_share_client(file_share_name)
-
-    # Create a directory (folder)
-    directory_client = file_share_client.get_directory_client(folder_name)
-    directory_client.create_directory()
-
-    print(f"Folder '{folder_name}' created successfully in Azure File Share!")
-
-
-def create_project_in_file_share(folder_name) -> None:
+class AzureFileShareClient:
     '''
-    ## Creates a Default Structured Project Folder in the Azure File Share Cloud Storage
-    ### Folder Structure: projects/project_name/folder_name
-    #### Currently only creates 3 folders: Folder1, Folder2, Folder3
+    ### Creates a client class for the Azure File Share Cloud Storage
+    You must set *AZURE_FILE_SHARE_CONNECTION_STRING* and *AZURE_FILE_SHARE_NAME*, which come from settings.py  
+    #### Currently can do the following:
+    - Create Folder
+    - Create Subfolders
+    - Create Project (either empty or with template)
+
+    So far, you NEED to have two folders in your main file share directory: "projects" and "templates"
+
+    The templates folder should have a "default" folder filled with random stuff.
+
+    (I'll automate the creation of this stuff later...)
+
+    TODO:
+    - Create Template Folder
+    - Upload Files to Project Folder
     '''
+    def __init__(self) -> None:
+        self.BASE_PATH = "projects/"
+        self.DEFAULT_TEMPLATE_PATH = "templates/"
 
-    PARENT_FOLDER = "projects/"
-    SUB_FOLDER = PARENT_FOLDER + folder_name + "/"
-    DEFAULT_TEST_FOLDERS = ["Folder1", "Folder2", "Folder3"]
+        self.connection_string: str = settings.AZURE_FILE_SHARE_CONNECTION_STRING
+        self.share_service_client: ShareServiceClient = ShareServiceClient.from_connection_string(self.connection_string)
 
-    try:
-        # Get the share service client
-        share_service_client = get_share_service_client()
+        self.file_share_name: str = settings.AZURE_FILE_SHARE_NAME
+        self.file_share_client: ShareClient = self.share_service_client.get_share_client(self.file_share_name)
+    
+    def create_folder_directory(self, folder_path_name: str) -> ShareDirectoryClient:
+        '''
+        #### Creates an empty folder in the given folder name
+        
+        :param str folder_path_name: The path to the folder to create
+        :return: a ShareDirectoryClient for further operations
+        '''
+        directory_client = self.file_share_client.get_directory_client(self.BASE_PATH + folder_path_name)
 
-        # Get the file share client
-        file_share_name = settings.AZURE_FILE_SHARE_NAME
-        file_share_client = share_service_client.get_share_client(file_share_name)
-
-        parent_dir = file_share_client.get_directory_client(PARENT_FOLDER)
-
-        # Creates the parent directory if it doesn't exist
-        if not parent_dir.exists():
-            parent_dir.create_directory()
-
-        # Create a directory (folder)
-        directory_client = file_share_client.get_directory_client(PARENT_FOLDER + folder_name)
-        directory_client.create_directory()
-
-
-        # Add Subfolders to the folder
-        for file_name in DEFAULT_TEST_FOLDERS:
-            print(f"Creating file '{file_name}' in folder '{folder_name}'...")
-            directory_client = file_share_client.get_directory_client(SUB_FOLDER + file_name)
+        try:
             directory_client.create_directory()
+            print(f"Folder '{folder_path_name}' created successfully in Azure File Share!")
 
-        # Load from template
-        # template = load_from_template(folder_name)
+        except ResourceExistsError:
+            print(f"Folder '{folder_path_name}' already exists in this directory! Aborting...")
+        
+        except Exception as ex:
+            print(f"An error occurred while creating folder '{folder_path_name}' in Azure File Share: {ex}")
+            raise ex
 
-        print(f"Folder '{folder_name}' created successfully in Azure File Share!")
+        return directory_client
 
-    except ResourceExistsError as ex:
-        print(f"Error Creating Folder. Folder '{folder_name}' already exists: {ex}")
-        # If an error occurs during creation, delete any folders that were created to avoid any resource leaks
-
-    except Exception as ex:
-        print(f"Error Creating Folder: {ex}")
-
-def load_from_template(folder_name) -> ShareDirectoryClient:
-    '''
-    ## Loads a file from a template in the Azure File Share Cloud Storage
-    ### Folder Structure: templates/folder_name
-    @folder_name: The type of template you want to load (basic only)
-
-    TODO: 
-    - DOESN'T WORK YET
-    - Have a dynamic template loader
-    - Add more templates
-    '''
-    TEMPLATE_PATH = "templates/" + folder_name
-
-    try:
-        # Get the share service client
-        share_service_client = get_share_service_client()
-
-        # Get the file share client
-        file_share_name = settings.AZURE_FILE_SHARE_NAME
-        file_share_client = share_service_client.get_share_client(file_share_name)
-
-        # Retrive folder
-        directory_client = file_share_client.get_directory_client(TEMPLATE_PATH)
+    def create_sub_folder_directory(self, folder_path_name: str, sub_folder_name: str) -> ShareDirectoryClient:
+        '''
+        #### Creates an empty subfolder in the given folder name
+        
+        :param str folder_path_name: The path to the folder to create
+        :param str sub_folder_name: The name of the subfolder
+        :return: A ShareDirectoryClient for further operations
+        '''
+        directory_client = self.file_share_client.get_directory_client(self.BASE_PATH + folder_path_name)
 
         if not directory_client.exists():
-            raise Exception(f"Folder '{TEMPLATE_PATH}' does not exist in Azure File Share! This is required for this function to work.")
+            print(f"Folder {folder_path_name} does not exist! Aborting...")
+            return
+
+        try:
+            directory_client.create_subdirectory(sub_folder_name)
+
+        except ResourceExistsError:
+            print(f"SubFolder '{sub_folder_name}' already exists in this directory! Aborting...")
+
+        except Exception as ex:
+            print(f"An error occurred while creating SubFolder '{sub_folder_name}' in Azure File Share: {ex}")
+            raise ex
         
         return directory_client
 
-    except Exception as ex:
-        RuntimeError(f"Error Loading Template Folder: {ex}")
+    def upload_file(self, folder_path_name: str, file_name: str, file_path: str):
+        '''CURRENTLY NOT TESTED.'''
+        directory_client = self.file_share_client.get_directory_client(folder_path_name)
+        directory_client.upload_file(file_path, file_name)
 
-def create_template_folder(folder_name) -> None:
-    '''
-    ## Creates a Template Folder in the Azure File Share Cloud Storage
-    ### Folder Structure: templates/folder_name
-    @folder_name: The type of template you want to load (basic only)
+        print(f"File '{file_name}' uploaded successfully to folder '{folder_path_name}' in Azure File Share!")
 
-    TODO:
-       - Not working
-    '''
+    def create_empty_project_folder(self, folder_path_name: str) -> None:
+        self.create_folder_directory(folder_path_name)
 
-    TEMPLATE_PATH = "templates/" + folder_name
+        print(f"Folder '{folder_path_name}' created successfully in Azure File Share!")
 
-    # DEBUG: Creating text files
-    text_file = open("test.txt", "x+")
-    text_file.write("This is a test file")
+        self.create_sub_folder_directory(folder_path_name, "Folder1")
+        self.create_sub_folder_directory(folder_path_name, "Folder2")
+        self.create_sub_folder_directory(folder_path_name, "Folder3")
 
-    try:
-        # Get the share service client
-        share_service_client = get_share_service_client()
+    def create_template_project_folder(self, folder_path_name: str, template_name: str) -> None:
+        '''
+        #### Creates a new project folder from a template
 
-        # Get the file share client
-        file_share_name = settings.AZURE_FILE_SHARE_NAME
-        file_share_client = share_service_client.get_share_client(file_share_name)
+        :param str folder_path_name: 
+            The path to the parent folder to create
+        :param str template_name:
+            The name of the template to use
+        '''
+        template_dir = self.file_share_client.get_directory_client(self.DEFAULT_TEMPLATE_PATH + template_name + "/")
+        dest_folder = self.create_folder_directory(folder_path_name)
 
-        # Create a directory (folder)
-        directory_client = file_share_client.get_directory_client(TEMPLATE_PATH)
+        self.recursively_create_template_project_folder(folder_path_name, template_dir, dest_folder)
 
-        if not directory_client.exists():
-            print(f"Creating folder '{TEMPLATE_PATH}' in Azure File Share...")
-            directory_client.create_directory()
+    def recursively_create_template_project_folder(self, folder_path_name: str, template_dir: ShareDirectoryClient, dest_folder: ShareDirectoryClient) -> None:
+        files = template_dir.list_directories_and_files()
         
-        # Adding a file
-        directory_client.upload_file("test.txt")
+        # Base Case
+        if not files:
+            return
 
-        print(f"Folder '{folder_name}' created successfully in Azure File Share!")
+        for file in files:
+            if file.is_directory:
+                dest_sub_folder = dest_folder.create_subdirectory(file.name)
+                template_sub_dir = template_dir.get_subdirectory_client(file.name)
 
-    except ResourceExistsError as ex:
-        print(f"Error Creating Folder. Folder '{folder_name}' already exists: {ex}")
+                # Recurisvely gather and create subfolders
+                self.recursively_create_template_project_folder(folder_path_name + "/" + file.name, template_sub_dir, dest_sub_folder)
+            else:
+                template_file = template_dir.get_file_client(file.name)
+                dest_file = dest_folder.get_file_client(file.name)
 
-    except Exception as ex:
-        print(f"Error Creating Template Folder: {ex}")
-
-    finally:
-        # Remove the created text file [DEBUG ONLY]
-        text_file.close()
-        os.remove("test.txt")
-    
-def add_files_to_folder(folder_name, file_name, file_path):
-    '''
-    ## Adds a file(s) to an existing folder in the Azure File Share Cloud Storage
-
-    TODO: this later
-    '''
-    pass
+                dest_file.start_copy_from_url(template_file.url)
