+++ backend/mysite/Trinity_Project/views.py
@@ -7,7 +7,7 @@ from rest_framework import status
 from rest_framework.exceptions import AuthenticationFailed
 from rest_framework.permissions import IsAuthenticated
 
-from .azure_file_share import create_folder_in_file_share, create_project_in_file_share
+from .azure_file_share import AzureFileShareClient
 from .models import Project
 from .models import User
 from .serializers import ProjectSerializer, UserNameSerializer, UserSerializer
@@ -104,7 +104,11 @@ def project_list(request):
             serializer.save()
             print("Serializer errors:", serializer.errors) # Debug
 
-            create_project_in_file_share(serializer.data['folder_location']) # Creates the test folders in the file share
+            folder = AzureFileShareClient()
+            if serializer.data['template'] == '':
+                folder.create_empty_project_folder(serializer.data['folder_location'])
+            else:
+                folder.create_template_project_folder(serializer.data['folder_location'], serializer.data['template'])
 
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -179,6 +183,6 @@ def create_azure_file_share_folder_view(request):
     #print(folder_name) 
     print(request.data) # This WILL print out the request body as an Object
     
-    create_folder_in_file_share(request.data['folder_name']) # Request only needs one field, folder_name, maybe later, we can specify it's location
+    # create_folder_in_file_share(request.data['folder_name']) # Request only needs one field, folder_name, maybe later, we can specify it's location
     
     return JsonResponse({'message': f'Folder "{folder_name}" created successfully in Azure File Share!'})
